requirements.txt
Flask==3.0.2
gunicorn==23.0.0
mysql-connector-python==8.3.0
numpy==1.26.4
pandas==2.2.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-experimental==0.3.4
langchain-huggingface==0.2.0
langchain-openai==0.3.18
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.3.42
openai==1.82.0
pinecone-client==5.0.1
python-dotenv==1.0.0 